app/backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "qr-backend"
version = "0.1.0"
description = "QR lifecycle backend"
authors = ["QR Team <dev@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
SQLAlchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
Pillow = "^10.2.0"
qrcode = "^7.4.2"
redis = "^5.0.0"
pydantic = "^1.10.7"
alembic = "^1.12.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.0"
ruff = "^0.1.5"
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
